python/requirements.txt
Flask>=2.0.1
Flask-SQLAlchemy>=3.0.2
PyMySQL>=1.0.2
python-dotenv>=0.19.0
yt-dlp>=2025.01.26